asserts: move base declaration tests
